TIP-221: Move ProductBuilder
