+++ CHANGELOG-1.6.md
@@ -13,6 +13,7 @@
 
 ## BC breaks
 
+- Move `Pim\Bundle\CatalogBundle\Builder\ProductBuilder` to `Pim\Component\Catalog\Builder\ProductBuilder`
 - Move `Pim\Bundle\CatalogBundle\Manager\AttributeValuesResolver` to `Pim\Component\Catalog\Manager\AttributeValuesResolver`
 - Remove deprecated `Pim\Bundle\CatalogBundle\Manager\FamilyManager`.
 - Remove deprecated `Pim\Bundle\CatalogBundle\Manager\ProductManager`.
