+++ src/Pim/Bundle/CatalogBundle/Builder/ProductTemplateBuilder.php
@@ -2,6 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Builder;
 
+use Pim\Component\Catalog\Builder\ProductBuilderInterface;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\ProductTemplateInterface;
 use Pim\Component\Catalog\Model\ProductValueInterface;
@@ -24,7 +25,7 @@ class ProductTemplateBuilder implements ProductTemplateBuilderInterface
     /** @var DenormalizerInterface */
     protected $denormalizer;
 
-    /** @var ProductBuilder */
+    /** @var ProductBuilderInterface */
     protected $productBuilder;
 
     /** @var LocaleResolver */
@@ -37,17 +38,17 @@ class ProductTemplateBuilder implements ProductTemplateBuilderInterface
     protected $productClass;
 
     /**
-     * @param NormalizerInterface   $normalizer
-     * @param DenormalizerInterface $denormalizer
-     * @param ProductBuilder        $productBuilder
-     * @param LocaleResolver        $localeResolver
-     * @param string                $productTemplateClass
-     * @param string                $productClass
+     * @param NormalizerInterface     $normalizer
+     * @param DenormalizerInterface   $denormalizer
+     * @param ProductBuilderInterface $productBuilder
+     * @param LocaleResolver          $localeResolver
+     * @param string                  $productTemplateClass
+     * @param string                  $productClass
      */
     public function __construct(
         NormalizerInterface $normalizer,
         DenormalizerInterface $denormalizer,
-        ProductBuilder $productBuilder,
+        ProductBuilderInterface $productBuilder,
         LocaleResolver $localeResolver,
         $productTemplateClass,
         $productClass
